+++ src/index.d.ts
@@ -435,6 +435,20 @@ export declare function asyncFilter<T> (
   func: (item: T) => MaybePromise<boolean>,
   iterable: AsyncIterableLike<T>
 ): AsyncIterableIterator<T>
+export declare function asyncFilter<S extends T, T> (concurrency: number, func: (item: T) => item is S):
+  (iterable: AsyncIterableLike<T>) => AsyncIterableIterator<S>
+export declare function asyncFilter<S extends T, T> (
+  concurrency: number,
+  func: (item: T) => item is S,
+  iterable: AsyncIterableLike<T>
+): AsyncIterableIterator<S>
+export declare function asyncFilter<T> (concurrency: number, func: (item: T) => MaybePromise<boolean>):
+  (iterable: AsyncIterableLike<T>) => AsyncIterableIterator<T>
+export declare function asyncFilter<T> (
+  concurrency: number,
+  func: (item: T) => MaybePromise<boolean>,
+  iterable: AsyncIterableLike<T>
+): AsyncIterableIterator<T>
 
 export declare function asyncFind<S extends T, T> (func: (item: T) => item is S):
   (iterable: AsyncIterableLike<T>) => Promise<S | null>
@@ -497,6 +511,13 @@ export declare function asyncIterable<T> (
 export declare function asyncMap<T, O> (func: (item: T) => MaybePromise<O>):
   (iter: AsyncIterableLike<T>) => AsyncIterableIterator<O>
 export declare function asyncMap<T, O> (
+  func: (item: T) => MaybePromise<O>,
+  iter: AsyncIterableLike<T>
+): AsyncIterableIterator<O>
+export declare function asyncMap<T, O> (concurrency: number, func: (item: T) => MaybePromise<O>):
+  (iter: AsyncIterableLike<T>) => AsyncIterableIterator<O>
+export declare function asyncMap<T, O> (
+  concurrency: number,
   func: (item: T) => MaybePromise<O>,
   iter: AsyncIterableLike<T>
 ): AsyncIterableIterator<O>